Implement `deployr <repo>` command
This implements the deployment workflow
  - git clone
  - docker build
  - docker push
  - kubectl apply

Point the `deployr` tool at a git repository and it will build
and deploy the application as a kubernetes pod.

So far, the implementation assumes a localhost kubernetes instance,
and a localhost docker registry. Also, the status information
assumes a single pod for simplicity. Later on, we can switch to
using kubernetes Deployments which allows us to track rolling
updates, horizontal scaling and better support with Services,
which let us assign a virtual IP/DNS to a collection of pods.

Simplifying with a single Pod means that service dependencies
can be managed on a per-app basis, agnostic to this tool.

This means that databases and queues can be included inside the Pod
(either directly, or as a sidecar proxy to a real provider).

Notes:
  `deployr <repo> <git hash>`

is not yet implemented
# Conflicts:
#	README