Document how to run multiple instances of a webapp
Also, when doing local development, it is useful to use a local
copy of the application being deployed. If `deployr` does not
correctly detect the first argument as a git repository, you can
force the tool to use the `deploy` subcommand.

We'll be adding more subcommands, e.g. `deployr info`, but `deploy`
is the most common, so we'll default to using that.